Fixed link to downloads section
The previous version of this link is interpreted relative to README.md access and does not stand alone if the source is downloaded. Changing this to an absolute link and pointing to the right location (instead of resulting in a Github 404 URI) will help users get the binary files they desire.